refactor(feature): split pipeline hydration support
